Moved the DiscussionServer in x_module.py, to be available in studio
